tramplin-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
geoalchemy2==0.14.1
httpx==0.25.1
redis==5.0.1
celery==5.3.4